refactor(language-core): do not generate `__VLS_types.d.ts` virtual file
